roles Jurado y Tutor integrados, añadidos a la barra de navegación
